+++ addOns/spider/CHANGELOG.md
@@ -5,7 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## Unreleased
-
+### Changed
+- Handle multiple "Link" HTTP Response headers.
+- Maintenance changes.
 
 ## [0.7.0] - 2023-10-12
 ### Changed